wiki_dataset_builder/requirements.txt
datasets>=2.14.0
beautifulsoup4>=4.12.0
pyarrow>=12.0.0
tqdm>=4.65.0
loguru>=0.7.0
pyyaml>=6.0
psutil>=5.9.0